feat(react): move package back to experimental
